+++ src/main/java/me/noodles/nv/NightVision.java
@@ -1,0 +1,68 @@
+package me.noodles.nv;
+
+import org.bukkit.Bukkit;
+import org.bukkit.command.CommandExecutor;
+import org.bukkit.event.Listener;
+import org.bukkit.plugin.Plugin;
+import org.bukkit.plugin.PluginDescriptionFile;
+import org.bukkit.plugin.java.JavaPlugin;
+
+public final class NightVision extends JavaPlugin implements Listener {
+
+
+    public static NightVision plugin;
+    public static Plugin plugin2;
+    private UpdateChecker checker;
+    String DownloaderType;
+    String GsonEntry;
+    String SpigotHost;
+
+
+
+    public void onEnable() {
+        NightVision.plugin = this;
+        final PluginDescriptionFile VarUtilType = this.getDescription();
+        this.getLogger().info("NightVision  V" + VarUtilType.getVersion() + " starting...");
+        this.saveDefaultConfig();
+        this.reloadConfig();
+        registerEvents((Plugin)this, new UpdateJoinEvent());
+        registerEvents(this, this);
+        this.getCommand("nv").setExecutor((CommandExecutor)new CommandNV());
+        this.getLogger().info("NightVision  V" + VarUtilType.getVersion() + " started!");
+        this.setEnabled(true);
+        this.getLogger().info("NightVision V" + VarUtilType.getVersion() + " checking for updates...");
+        this.checker = new UpdateChecker(this);
+        if (this.checker.isConnected()) {
+            if (this.checker.hasUpdate()) {
+                getServer().getConsoleSender().sendMessage("------------------------");
+                getServer().getConsoleSender().sendMessage("NightVision is outdated!");
+                getServer().getConsoleSender().sendMessage("Newest version: " + this.checker.getLatestVersion());
+                getServer().getConsoleSender().sendMessage("Your version: " + NightVision.plugin.getDescription().getVersion());
+                getServer().getConsoleSender().sendMessage("Please Update Here: https://www.spigotmc.org/resources/46693");
+                getServer().getConsoleSender().sendMessage("------------------------");
+            }
+            else {
+                getServer().getConsoleSender().sendMessage("------------------------");
+                getServer().getConsoleSender().sendMessage("NightVision is up to date!");
+                getServer().getConsoleSender().sendMessage("------------------------");            }
+        }
+    }
+
+    public static void registerEvents(final Plugin plugin, final Listener... listeners) {
+        for (final Listener listener : listeners) {
+            Bukkit.getServer().getPluginManager().registerEvents(listener, plugin);
+        }
+    }
+
+    @SuppressWarnings({ "unchecked", "rawtypes"})
+    public static NightVision getPlugin() {
+        return (NightVision)getPlugin((Class)NightVision.class);
+    }
+
+    public static Plugin getPlugin2() {
+        return (Plugin)NightVision.plugin;
+    }
+
+
+}
+
