+++ src/main/resources/plugin.yml
@@ -1,0 +1,8 @@
+name: NightVision
+version: ${project.version}
+main: me.noodles.nv.NightVision
+authors: [BGHDDevelopment]
+website: http://bghddevelopment.com/
+commands:
+  nv:
+    description: Enable or Disable NV
